% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_tab.R
\name{all_summary}
\alias{all_summary}
\title{Summarize a vector}
\usage{
all_summary(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list of summary statistics.
}
\description{
\code{all_summary()} provide a vector of summary statistics including.
\itemize{
\item Minimum
\item Median
\item Mean
\item Maximum
\item Standard deviation
\item Inter quarterly range
\item Sample size
}
}
\examples{
all_summary(iris$Sepal.Length)

}
\seealso{
\code{\link[=summary]{summary()}} for base summary statistics.
}
