% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_pack.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_by}
\alias{mutate_if}
\alias{group_map}
\alias{enquo}
\alias{as_label}
\alias{IQR}
\alias{median}
\alias{sd}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_map}}, \code{\link[dplyr:mutate_all]{mutate_if}}}

  \item{rlang}{\code{\link[rlang]{as_label}}, \code{\link[rlang]{enquo}}}

  \item{stats}{\code{\link[stats]{IQR}}, \code{\link[stats]{median}}, \code{\link[stats]{sd}}}
}}

