% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_tab.R
\name{sum_tab}
\alias{sum_tab}
\title{Summary statistics for a dataframe}
\usage{
sum_tab(.data, .class)
}
\arguments{
\item{.data}{A dataframe}

\item{.class}{Column name for group wise summary statistics. Default is  \code{NULL}.}
}
\value{
A dataframe or a list of dataframe of summary statistics.
}
\description{
\code{sum_tab()} create a data frame of summary statistics for all numeric column in a
dataframe. For grouping data, \code{sum_tab()} create a list of summary statistics
dataframe for each of the group.
}
\examples{
sum_tab(iris, Species)

}
